feat: add search feature
