+++ index.html
@@ -3,6 +3,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" integrity="sha384-GJzZqFGwb1QTTN6wy59ffF1BuGJpLSa9DkKMp0DgiMDm4iYMj70gZWKYbI706tWS" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
     <link rel="stylesheet" href="./style.css">
     <title>SW-App</title>
 </head>
@@ -35,18 +36,21 @@
                 <a class="nav-link active" href="#">Starships</a>
             </li>
           </ul>
-          <!-- SEARCH -->
-          <!-- <input placeholder="Search" type="text" class="form-control" id="search"> -->
+          <form class="search-form" action="action_page.php">
+            <input class="search-field" type="text" placeholder="Search.." name="search">
+            <button id="search"><i class="fa fa-search"></i></button>
+          </form>
         </div>
         <div class="starwars-demo">
           <img src="//cssanimation.rocks/demo/starwars/images/star.svg" alt="Star" class="star">
           <img src="//cssanimation.rocks/demo/starwars/images/wars.svg" alt="Wars" class="wars">
         </div> 
+        <div class="lds-facebook"><div></div><div></div><div></div></div>
         <div class="card-body">
         </div>  
       </div>    
     </div>
-      
+         
 
     <script src="./dist/bundle.js"></script>
 </body>
