+++ src/index.js
@@ -1,8 +1,14 @@
+import {loadPeople, loadPlanets, loadFilms, loadSpecies, loadVehicles} from './loadData.js';
+
+let category;
+let categorySelected = false;
 const cardsField = document.querySelector('.card-body');
 const categories = document.querySelectorAll(`.nav-link`);
 const paginator = document.createElement('ul');
-const loader = document.querySelector('.starwars-demo')
-let category;
+const introImage = document.querySelector('.starwars-demo');
+const loader = document.querySelector('.lds-facebook');
+const search = document.getElementById('search');
+const searchField = document.querySelector('.search-field');
 
 categories.forEach(el => {
     el.addEventListener('click', (e) => {
@@ -16,10 +22,26 @@ categories.forEach(el => {
     });
 })
 
+function getSearchData(category, value) {
+    cardsField.innerHTML = '';
+    fetch(`https://swapi.dev/api/${category.toLowerCase()}/?search=${value}`)
+    .then(res => res.json())
+    .then(data => {
+        if (data.results.length === 0) {
+            showMessage(`Don’t blame me. I’m an interpreter. I’m not supposed to know every ${category.toLowerCase()} in the Galaxy .`)
+        }
+       
+        generateCards(data.results, category)
+    })
+    .catch(err => console.log(err));
+}
+
 
 function loadCategory(category, page) {
     cardsField.innerHTML = '';
-    loader.style.display = 'flex';
+    introImage.style.display = 'none';
+    loader.style.display = 'inline-block';
+    categorySelected = true;
     fetch(`https://www.swapi.tech/api/${category}?page=${page}&limit=10`)
         .then(res => res.json())
         .then(data => loadDetails(data, category, page))
@@ -79,92 +101,9 @@ function generateCards(info, category) {
             default:
                 break;
             }
-        }  
-}
-
-
-function loadName(card, info, i){
-    let name = document.createElement('h5');
-    name.className = 'card-title'
-    name.innerText = info[i].name;
-    card.appendChild(name);
-}
-
-function loadTitle(card, info, i){
-    let name = document.createElement('h5');
-    name.className = 'card-title'
-    name.innerText = info[i].title;
-    card.appendChild(name);
-}
-
-
-
-
-function loadPeople(info, card, i) {
-    loadName(card, info, i);
-    loadProperties(card, "Gender: ", info, 'gender', i);
-    loadProperties(card, "Height: ", info, 'height', i);
-    loadProperties(card, "Mass: ", info, 'mass', i);
-    loadProperties(card, "Hair: ", info, 'hair_color', i);
-    loadProperties(card, "Skin: ", info, 'skin_color', i);
-    loadProperties(card, "Eye: ", info, 'eye_color', i);
-    loadProperties(card, "Birth: ", info, 'birth_year', i);
+        } 
 
 }
-
-
-function loadPlanets(info, card, i) {
-    loadName(card, info, i);
-    loadProperties(card, "Diameter: ", info, 'diameter', i);
-    loadProperties(card, "Gravity: ", info, 'gravity', i);
-    loadProperties(card, "Orbital period: ", info, 'orbital_period', i);
-    loadProperties(card, "Population: ", info, 'population', i);
-    loadProperties(card, "Rotation period: ", info, 'rotation_period', i);
-    loadProperties(card, "Terrain: ", info, 'terrain', i);
-    loadProperties(card, "Climate: ", info, 'climate', i);
-}
-
-function loadFilms(info, card, i) {
-    loadTitle(card, info, i);
-    loadProperties(card, "Episode: ", info, 'episode_id', i);
-    loadProperties(card, "Director: ", info, 'director', i);
-    loadProperties(card, "Opening crawl: ", info, 'opening_crawl', i);
-    loadProperties(card, "Release date: ", info, 'release_date', i);
-}
-
-function loadSpecies(info, card, i) {
-    loadName(card, info, i);
-    loadProperties(card, "Average height: ", info, 'average_height', i);
-    loadProperties(card, "Average lifespan: ", info, 'average_lifespan', i);
-    loadProperties(card, "Classification: ", info, 'classification', i);
-    loadProperties(card, "Eye colors: ", info, 'eye_colors', i);
-    loadProperties(card, "Hair colors: ", info, 'hair_colors', i);
-    loadProperties(card, "Skin colors: ", info, 'skin_colors', i);
-    loadProperties(card, "Language: ", info, 'language', i);
-}
-
-function loadVehicles(info, card, i) {
-    loadName(card, info, i);
-    loadProperties(card, "Length: ", info, 'length', i);
-    loadProperties(card, "Cargo capacity: ", info, 'cargo_capacity', i);
-    loadProperties(card, "Cost in credits: ", info, 'cost_in_credits', i);
-    loadProperties(card, "Crew: ", info, '46', i);
-    loadProperties(card, "Max atmosphering speed: ", info, 'max_atmosphering_speed', i);
-    loadProperties(card, "Model: ", info, 'model', i);
-    loadProperties(card, "Manufacturer: ", info, 'manufacturer', i);
-}
-
-function loadProperties(card, captionText, info, key, i){
-    let caption = document.createElement('p');
-    caption.innerText = captionText;
-    card.appendChild(caption);
-    let value = document.createElement('span');
-    value.className = 'badge badge-secondary';
-    value.innerText = info[i][key]
-    caption.appendChild(value);
-
-}
-
 
 
 function generatePagination(pages, page) { 
@@ -196,14 +135,32 @@ function moveToPage(e){
 }
 
 function notifyByError(err) {
-    if(err = 429){
-        let cardsContainer = document.createElement('div');
-        cardsContainer.className = 'card-columns';
-        cardsField.appendChild(cardsContainer);
+    console.log(err);
+    showMessage('Data is in a galaxy far, far away. Please, try later...');
 
-        let message = document.createElement('h5');
-        message.className = 'display-4';
-        message.innerText = 'Data is in a galaxy far, far away. Please, try later...';
-        cardsContainer.appendChild(message);
+}
+
+search.addEventListener('click', searchValue);
+
+function searchValue(e) {
+    e.preventDefault();
+    let input = searchField.value;
+    console.log(input)
+    if (!categorySelected) {
+        showMessage('Please, select category first');
+    } else {
+      
+        getSearchData(category, input);
+       
     }
+}
+
+function showMessage(text) {
+        cardsField.innerHTML = '';
+        introImage.style.display = 'none';
+        let message = document.createElement('h5');
+        message.className = 'message display-4';
+        message.innerText = text;
+        cardsField.appendChild(message);
+
 }
